KEA-108 move groupconstructor
